requirements.txt
beautifulsoup4==4.13.3
langchain==0.3.23
langchain_google_genai==2.1.2
python-dotenv==1.1.0
selenium==4.31.0
streamlit==1.32.0
webdriver_manager==4.0.2
